// Test11/src/models/profile.ts
export interface Profile {//new for profile
   
    firstName: string;
    lastName: string; 
    age: string;
    adress: string;
    phoneNumber: string;
    level: string;
    identification: string;
    establishment: string;
    group: {
            name:string;
            
           
            
           }
        
    
}